v2.1.2
* Addressed issue #8 - If no files are found, nothing is logged and the
script terminates. This cuts down on log clutter, as well as frees up
the CPU and process for those running the script very frequently as a
scheduled task.
* If $appendLog = $True, the script version header will only be printed
to the log if the log is empty in order to cut down on log clutter.
* "New session" header is only printed to the log if files are found and
added to the queue to cut down on log clutter.